requirements.txt
Django==4.2
djangorestframework==3.15.1
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-filter==24.1
